+++ CHANGELOG.md
@@ -65,6 +65,7 @@ in the detailed section referring to by linking pull requests or issues.
 * Add query validation (#1605)
 * Adapt logs to the logging guide (#1425)
 * Fix incompatibility `DecentralizedIdentityServiceExtension` and `FsPrivateKeyResolver` (#1696)
+* Fix federated catalog showing empty between clear and crawl execution completion (#1703)
 
 ## [milestone-4] - 2022-06-07